+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- fixed: Resyncing FIO wallets lose stake/unstake tx data
 - fixed: (Solana) Properly validate custom token addresses
 
 ## 4.22.0 (2024-08-30)